weebhook mercadopago, ticket, logica de compra

+++ Back-End/src/main/java/com/tiendropa/Tienda/de/Ropa/models/Orden.java
@@ -28,7 +28,7 @@ public class Orden {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private BigDecimal precioTotal;
+    private double precioTotal;
 
     private String metodoDePago;
